+++ webapp/shared/datex2/v3_5_json_static/models/energy_infrastructure_station_input.py
@@ -61,7 +61,7 @@ class EnergyInfrastructureStationInput(ValidataclassMixin):
         ListValidator(DataclassValidator(MultilingualStringInput)),
         Default(UnsetValue),
     )
-    totalMaximumPower: int = FloatValidator()
+    totalMaximumPower: float = FloatValidator()
     authenticationAndIdentificationMethods: list[AuthenticationAndIdentificationEnumGInput] | UnsetValueType = (
         ListValidator(DataclassValidator(AuthenticationAndIdentificationEnumGInput)),
         Default(UnsetValue),
