+++ webapp/shared/datex2/v3_5_json_static/models/length_characteristic_input.py
@@ -18,7 +18,7 @@ class LengthCharacteristicInput(ValidataclassMixin):
     """
 
     comparisonOperator: ComparisonOperatorEnumGInput = DataclassValidator(ComparisonOperatorEnumGInput)
-    vehicleLength: int = FloatValidator()
+    vehicleLength: float = FloatValidator()
     comLengthCharacteristicExtensionG: ExtensionTypeGInput | UnsetValueType = (
         DataclassValidator(ExtensionTypeGInput),
         Default(UnsetValue),
