+++ webapp/shared/datex2/v3_7_json_static/models/connector_input.py
@@ -28,9 +28,9 @@ class ConnectorInput(ValidataclassMixin):
         DataclassValidator(ConnectorFormatTypeEnumGInput),
         Default(UnsetValue),
     )
-    maxPowerAtSocket: int = FloatValidator()
-    voltage: int | UnsetValueType = FloatValidator(), Default(UnsetValue)
-    maximumCurrent: int | UnsetValueType = FloatValidator(), Default(UnsetValue)
+    maxPowerAtSocket: float = FloatValidator()
+    voltage: float | UnsetValueType = FloatValidator(), Default(UnsetValue)
+    maximumCurrent: float | UnsetValueType = FloatValidator(), Default(UnsetValue)
     egiConnectorExtensionG: ExtensionTypeGInput | UnsetValueType = (
         DataclassValidator(ExtensionTypeGInput),
         Default(UnsetValue),
