+++ webapp/shared/datex2/v3_7_json_static/models/dangerous_goods_extended_input.py
@@ -25,7 +25,7 @@ class DangerousGoodsExtendedInput(ValidataclassMixin):
     """
 
     chemicalName: MultilingualStringInput = DataclassValidator(MultilingualStringInput)
-    dangerousGoodsFlashPoint: int | UnsetValueType = FloatValidator(), Default(UnsetValue)
+    dangerousGoodsFlashPoint: float | UnsetValueType = FloatValidator(), Default(UnsetValue)
     dangerousGoodsRegulations: DangerousGoodsRegulationsEnumGInput | UnsetValueType = (
         DataclassValidator(DangerousGoodsRegulationsEnumGInput),
         Default(UnsetValue),
@@ -35,8 +35,8 @@ class DangerousGoodsExtendedInput(ValidataclassMixin):
     hazardSubstanceItemPageNumber: str | UnsetValueType = StringValidator(), Default(UnsetValue)
     tremCardNumber: str | UnsetValueType = StringValidator(), Default(UnsetValue)
     undgNumber: str | UnsetValueType = StringValidator(), Default(UnsetValue)
-    volumeOfDangerousGoods: int | UnsetValueType = FloatValidator(), Default(UnsetValue)
-    weightOfDangerousGoods: int | UnsetValueType = FloatValidator(), Default(UnsetValue)
+    volumeOfDangerousGoods: float | UnsetValueType = FloatValidator(), Default(UnsetValue)
+    weightOfDangerousGoods: float | UnsetValueType = FloatValidator(), Default(UnsetValue)
     adrClassValue: list[str] | UnsetValueType = ListValidator(StringValidator()), Default(UnsetValue)
     comHazardousMaterialsExtensionG: ExtensionTypeGInput | UnsetValueType = (
         DataclassValidator(ExtensionTypeGInput),
