+++ webapp/shared/datex2/v3_7_json_static/models/floating_point_metre_distance_value_input.py
@@ -16,7 +16,7 @@ class FloatingPointMetreDistanceValueInput(ValidataclassMixin):
     A measured or calculated value of distance in metres in a floating point format.
     """
 
-    distance: int = FloatValidator()
+    distance: float = FloatValidator()
     comDataValueExtensionG: ExtensionTypeGInput | UnsetValueType = (
         DataclassValidator(ExtensionTypeGInput),
         Default(UnsetValue),
