+++ webapp/shared/datex2/v3_7_json_static/models/height_coordinate_input.py
@@ -17,7 +17,7 @@ class HeightCoordinateInput(ValidataclassMixin):
     Third coordinate for points defined geodetically
     """
 
-    heightValue: int = FloatValidator()
+    heightValue: float = FloatValidator()
     heightType: HeightTypeEnumGInput | UnsetValueType = DataclassValidator(HeightTypeEnumGInput), Default(UnsetValue)
     locHeightCoordinateExtensionG: ExtensionTypeGInput | UnsetValueType = (
         DataclassValidator(ExtensionTypeGInput),
