+++ webapp/shared/datex2/v3_7_json_static/models/precipitation_intensity_value_input.py
@@ -16,7 +16,7 @@ class PrecipitationIntensityValueInput(ValidataclassMixin):
     A measured or calculated value of the accumulation rate of precipitation.
     """
 
-    millimetresPerHourIntensity: int = FloatValidator()
+    millimetresPerHourIntensity: float = FloatValidator()
     comDataValueExtensionG: ExtensionTypeGInput | UnsetValueType = (
         DataclassValidator(ExtensionTypeGInput),
         Default(UnsetValue),
