+++ webapp/shared/datex2/v3_7_json_static/models/temperature_below_or_above_road_surface_input.py
@@ -17,7 +17,7 @@ class TemperatureBelowOrAboveRoadSurfaceInput(ValidataclassMixin):
     Mesurement of temperature below or above the road surface.
     """
 
-    heightBelowOrAboveRoadSurface: int = FloatValidator()
+    heightBelowOrAboveRoadSurface: float = FloatValidator()
     temperatureBelowOrAboveRoadSurface: TemperatureValueInput = DataclassValidator(TemperatureValueInput)
     comTemperatureBelowOrAboveRoadSurfaceExtensionG: ExtensionTypeGInput | UnsetValueType = (
         DataclassValidator(ExtensionTypeGInput),
