+++ webapp/shared/datex2/v3_7_json_static/models/wind_speed_value_input.py
@@ -16,7 +16,7 @@ class WindSpeedValueInput(ValidataclassMixin):
     A measured or calculated value of wind speed.
     """
 
-    windSpeed: int = FloatValidator()
+    windSpeed: float = FloatValidator()
     comDataValueExtensionG: ExtensionTypeGInput | UnsetValueType = (
         DataclassValidator(ExtensionTypeGInput),
         Default(UnsetValue),
